# Make ssf.RandomlySample a million times faster