@ChimeraCoder 2018-04-25
Because RandomlySample mutates its data, the benchmark had a race condition in it, though not one which is likely to manifest in practice (I couldn't reproduce locally, but Andreas did, and either way, it clearly existed).

To solve the race condition, I'm allocating the samples fresh within each run. This significantly slows down the reported execution time, but mostly because we're now allocating a new slice and recreating the `Count` samples within each run. What we really care about is benchmarking the difference in implementations, so the benchmark still gives us meaningful data.

The original benchmark (with the rare race condition) is still probably the better estimator of the raw performance of RandomlySample in parallel, to the extent that we care about having an estimate of the raw performance in a hermetic environment. At the end of the day, we got the real-life profiles from production volume, and that's where it *really* matters. :)