@ChimeraCoder 2018-04-25
(@asf-stripe gave the thumbs-up to the changes via Slack.)